requirements.txt
tensorflow==2.16.1
keras==3.3.3
Flask==3.0.2
Flask_RESTful==0.3.10
Flask-Cors==1.10.3
pillow==10.3.0
scipy==1.13.0
